% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conctoquant.R
\name{conctoquant}
\alias{conctoquant}
\title{conctoquant}
\usage{
conctoquant(hypso.curve = data2, input, first.data.col = 3)
}
\arguments{
\item{hypso.curve}{(m-sqm) [as.double or as.integer] an hypsometric curve containig the area of each depth of the lake}

\item{input}{(-) input data as dataframe}

\item{first.data.col}{(-) the first column of data for interpolation}
}
\value{
quantified data for the lake according to the hypsometric curve
}
\description{
this function transforms concentrations into quantities using the hypsometric curve
}
\author{
Yoav BD
}

