% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datedifference.R
\name{datedifference}
\alias{datedifference}
\title{date difference}
\usage{
datedifference(input)
}
\arguments{
\item{input}{(-) input data as list of data frame splitted into data frames by years}
}
\value{
Date_diff a list of vactros including the difference between dates and the sequence of accumulated time between dates
}
\description{
this function extracts a segment of a data set according to a selected date range
}
\author{
Yoav BD
}

