% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datedifferencedf.R
\name{datedifferencedf}
\alias{datedifferencedf}
\title{date difference df}
\usage{
datedifferencedf(input)
}
\arguments{
\item{input}{(-) input data as list of data splitted into data frames by years}
}
\value{
Date_diff a list of vectros including the difference between dates and the sequence of accumulated time between dates
}
\description{
this function calculates the difference between measurments in days for caculation of time-dependent processes, and is adjusted for a dataframe (rather than datedifference wich is suitable for a year splitted list)
}
\author{
Yoav BD
}

