% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datesegmenter.R
\name{datesegmenter}
\alias{datesegmenter}
\title{date segmenter}
\usage{
datesegmenter(max.date, min.date, first.data.col = 3, input)
}
\arguments{
\item{max.date}{(-) the maximum range date to be extracted}

\item{min.date}{(-) the minimum range date to be extracted}

\item{first.data.col}{(-) the first column of data for interpolation}

\item{input}{(-) input data as dataframe}
}
\value{
only the data that falls into the required date range
}
\description{
this function extracts a segment of a data set according to a selected date range
}
\author{
Yoav BD
}

