% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datesplitter.R
\name{datesplitter}
\alias{datesplitter}
\title{date splitter}
\usage{
datesplitter(input, max.date, min.date)
}
\arguments{
\item{input}{(-) input data as dataframe}
}
\value{
only the data that falls into the required date range
}
\description{
this function splits the input data.frame into a list of data frames according to a specified date range
}
\author{
Yoav BD
}

