% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depthinterp.R
\name{depthinterp}
\alias{depthinterp}
\title{interpolate data by depth}
\usage{
depthinterp(maxdepth = 40, intreval = 1, first.data.col = 3, input,
  col.num = as.integer(ncol(i[[1]])), depth.col = 2)
}
\arguments{
\item{maxdepth}{(m) max depth to be interpolated for the lake}

\item{intreval}{(m) the depth difference between each of the interpolation points, default is 1 m}

\item{first.data.col}{(-) the first column of data for interpolation}

\item{input}{(-) input data as list}

\item{col.num}{(-) number of columns in the original data table}

\item{depth.col}{(-) the column that contains the depths of measurments}
}
\value{
interpolated data for the  (arbitrary units)
}
\description{
this function intepolates a list comprising the N measurments of a lake
package zoo is required to complete NA values in the matrix
}
\author{
Yoav BD
}

