% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitequation.R
\name{fitequation}
\alias{fitequation}
\title{fit equation}
\usage{
fitequation(input)
}
\arguments{
\item{input}{(-) input data as dataframe}
}
\value{
only the data that falls into the required date range
}
\description{
this function retrieves the equation for the a linear model fit
}
\author{
Yoav BD
}

