% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakemixer.R
\name{lakemixer}
\alias{lakemixer}
\title{lake mixer}
\usage{
lakemixer(input, max.depth = 40)
}
\arguments{
\item{input}{(-) input data as list which has a column with name Oxygen and a column named Depth}

\item{first.data.col}{(-) the first column of data to mix}
}
\value{
the geochemical properties of the mixed lake
}
\description{
this function simulate lake mixing
}
\author{
Yoav BD
}

