% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeNH4.R
\name{lakeNH4}
\alias{lakeNH4}
\title{lakeNH4}
\usage{
lakeNH4(model.data, mean.rand.NH4 = 3e+05, sd.rand.NH4 = 3e+05/10,
  data.dates = as.Date(unique(mode.data$Date)),
  mix.index = as.Date(data.dates[1]), mix.delay.ind = 0,
  first.mix = as.Date(as.Date(data.dates[length(data.dates)/2])),
  parameters.NH4.consumption, NH4.comsumption.correction.factor = 1,
  date.dif.for.model.seq, oxycline.depth.date, crit.depth = 35,
  mixing.ratio.NH4 = 0.95, parameters.NH4.buildup, dur2 = 30, dur3 = 120,
  first.mix.delay.ind = 1.5 * dur3, max.depth)
}
\arguments{
\item{model.data}{a data frame that has the first time step and is ready to incorporate all the results of the calculations. row number should equal the length of time series times the depths interpolated. the data.frame should include columns named Date, Depth, and NH4}

\item{mean.rand.NH4}{(max.depth) the mean value for the noise factor of the process}

\item{sd.rand.NH4}{(concentration) the sd value for the noise factor of the process}

\item{data.dates}{(date in official non-ambigous form "yyyy-mm-dd" ) the dates of the data to be evaluated}

\item{mix.index}{(date in official non-ambigous form "yyyy-mm-dd" )the last mix of the lake. default is first date of time series}

\item{mix.delay.ind}{(days) an integer indicating time since last mixing of the lake}

\item{first.mix}{(date in official non-ambigous form "yyyy-mm-dd" )the last mix of the lake. default is the middle date of the time series}

\item{parameters.NH4.consumption}{(integers) a set of parameters for the linear fit of NH4 consumption. number of coefficients should match interpolation depth}

\item{NH4.comsumption.correction.factor}{(integer) the mean value for the noise factor of the process}

\item{date.dif.for.model.seq}{(integers) a vector of integers indicating the time difference between each of the measurments}

\item{oxycline.depth.date}{(Date-depth) the depth of the oxycline in each of dates to be analyzed}

\item{crit.depth}{(m) the depth of the oxycline from which the lake is considered mixed}

\item{mixing.ratio.NH4}{(-) a value indicating a factor to multiply after mixing to aacount for consumption during mixing}

\item{parameters.NH4.buildup}{(integers) a set of parameters for the linear fit of NH4 buildup. number of coefficients should match interpolation depth}

\item{dur2}{(d) delay between first mixing of the seaon and the onset of NH4 consumption}

\item{dur3}{(d) maximum length of NH4 consumption period}

\item{first.mix.delay.ind}{(days) an integer indicating time since the first mixing of the lake in that year}

\item{max.depth}{(m) maximum depth to be interpolated}
}
\value{
only the data that falls into the required date range
}
\description{
This model carries the calculations of the NH4 cycle in the lake for a series of time-series data
}
\author{
Y. Ben Dor
}

