% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSE.R
\name{NSE}
\alias{NSE}
\title{NSE}
\usage{
NSE(m, o)
}
\arguments{
\item{m}{modelled value (any unit)}

\item{o}{observed value (any unit)}
}
\value{
only the data that falls into the required date range
}
\description{
Nash-Sutcliffe Efficiency Performance Metric
}
\author{
N. Tague
}

