% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realdepth.R
\name{realdepth}
\alias{realdepth}
\title{real depth}
\usage{
realdepth(depth.date = data3, measurment.date, input)
}
\arguments{
\item{depth.date}{(date-m) [as.Date-integer] a tabel (as a data frame) that has the water level of the measurment}

\item{measurment.date}{(date) [as.Date] the date during which the measurment was carried}

\item{input}{(-) input data as dataframe}
}
\value{
the data with a replacement of column of the realative depth with the real depths of each measurment
}
\description{
this function assigns the real depth of every measurment according to the measurment date
}
\author{
Yoav BD
}

