% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearassigner.R
\name{yearassigner}
\alias{yearassigner}
\title{year assigner}
\usage{
yearassigner(input)
}
\arguments{
\item{input}{(-) input data as dataframe}
}
\value{
the input data with an additional date formatted column indicating the year of the measurement
}
\description{
this function assigns the year of wach mesurment into a new column
}
\author{
Yoav BD
}

