% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearbyelement.R
\name{yearbyelement}
\alias{yearbyelement}
\title{year by element}
\usage{
yearbyelement(input)
}
\arguments{
\item{input}{(-) input data as list with element name to be used}
}
\value{
the input data as a list with the year value of each row according to the element
}
\description{
this function assigns the year value by the element name
}
\author{
Yoav BD
}

