% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datedifference.R
\name{datedifference}
\alias{datedifference}
\title{date difference}
\usage{
datedifference(input)
}
\arguments{
\item{input}{(-) input data as list of data splitted into data frames by years}
}
\value{
Date_diff a list of vactros including the difference between dates and the sequence of accumulated time between dates
}
\description{
this function calculates the difference between measurments in days for caculation of time-dependent processes
}
\author{
Yoav BD
}

